requirements.txt
alembic==0.9.2
aniso8601==1.2.1
appnope==0.1.0
click==6.7
decorator==4.0.11
factory-boy==2.8.1
Faker==0.7.15
fakeredis==0.8.2
Flask==0.12.2
Flask-JWT-Extended==2.2.0
Flask-Migrate==2.0.4
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
marshmallow==3.0.0b2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.34
Pygments==2.2.0
PyJWT==1.5.0
pytest==3.1.1
python-dateutil==2.6.0
python-dotenv==0.6.4
python-editor==1.0.3
pytz==2017.2
redis==2.10.5
simplegeneric==0.8.1
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.1.10
traitlets==4.3.2
wcwidth==0.1.7
webargs==1.7.0
Werkzeug==0.12.2
